performing selection filtering but check marks not visible in dropdowns
